requirements.txt
psycopg2-binary==2.9.9
pandas==2.2.2
openpyxl==3.1.2
tkcalendar==1.6.1
cryptography==41.0.7
pyinstaller==6.3.0